+++ app/src/main/java/de/baumann/browser/browser/NinjaWebChromeClient.java
@@ -74,7 +74,7 @@ public class NinjaWebChromeClient extends WebChromeClient {
             public boolean shouldOverrideUrlLoading(WebView view, WebResourceRequest request) {
                 //newWebView.loadUrl(request.getUrl().toString());
                 try {
-                    newWebView.loadUrl(request.getUrl().toString());
+                    BrowserUnit.intentURL(context, request.getUrl());
                 } catch (Exception e) {
                     Log.i(TAG, "shouldOverrideUrlLoading Exception:" + e);
                     Intent intent = new Intent(Intent.ACTION_VIEW);
@@ -116,12 +116,11 @@ public class NinjaWebChromeClient extends WebChromeClient {
     @Override
     public void onPermissionRequest(final PermissionRequest request) {
         SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(ninjaWebView.getContext());
-        String profile = sp.getString("profile", "profileStandard");
         Activity activity = (Activity) ninjaWebView.getContext();
         String[] resources = request.getResources();
         for (String resource : resources) {
             if (PermissionRequest.RESOURCE_VIDEO_CAPTURE.equals(resource)) {
-                if (sp.getBoolean(profile + "_camera", false)){
+                if (sp.getBoolean(NinjaWebView.getProfile() + "_camera", false)){
                     HelperUnit.grantPermissionsCamera(activity);
                     if (ninjaWebView.getSettings().getMediaPlaybackRequiresUserGesture())
                         ninjaWebView.getSettings().setMediaPlaybackRequiresUserGesture(false);
@@ -130,7 +129,7 @@ public class NinjaWebChromeClient extends WebChromeClient {
                     request.grant(request.getResources());
                 }
             } else if (PermissionRequest.RESOURCE_AUDIO_CAPTURE.equals(resource)) {
-                if (sp.getBoolean(profile + "_microphone", false)){
+                if (sp.getBoolean(NinjaWebView.getProfile() + "_microphone", false)){
                     HelperUnit.grantPermissionsMic(activity);
                     request.grant(request.getResources());
                 }